+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## vx.x.x
  - Use default audio output
  - Automatically find yowasp-yosys.exe binary on windows
+ - Use pathlib for examples and test
+ - Enable github/action for testing of python3.13 in windows
 
 ## v0.11.0
  - Fix settings load bug